requirements.txt
#anyio==4.4.0
#bidict==0.23.1
caninos-sdk==0.5.0
#click==8.1.7
#dnspython==2.6.1
eventlet==0.36.1
exceptiongroup==1.2.1
gpiod==2.1.3
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
Pygments==2.18.0
pyserial==3.5
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
simple-websocket==1.0.0
sniffio==1.3.1
typer==0.12.3
typing_extensions==4.11.0
uvicorn==0.30.1
#uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wsproto==1.2.0
